adicionando novos cadastros

+++ src/main/java/br/com/senac/curriculum/repository/educacao/EducacaoRepository.java
@@ -2,6 +2,6 @@ package br.com.senac.curriculum.repository.educacao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EducacaoRepository extends JpaRepository<EducacaoEntity, Integer> {
+public interface EducacaoRepository extends JpaRepository<EducacaoEntity, Long> {
 }
