+++ src/main/java/br/com/senac/curriculum/repository/endereco/EnderecoRepository.java
@@ -2,5 +2,5 @@ package br.com.senac.curriculum.repository.endereco;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoRepository extends JpaRepository<EnderecoEntity, Integer> {
+public interface EnderecoRepository extends JpaRepository<EnderecoEntity, Long> {
 }
